+++ templates/node/test/index.test.ts
@@ -1,3 +1,4 @@
+import type ParamsDictionary from "@nanoservice-ts/shared/dist/types/ParamsDictionary";
 import { beforeAll, expect, test } from "vitest";
 import Node from "../index";
 import ctx from "./helper";
@@ -12,5 +13,7 @@ beforeAll(() => {
 // Validate Hello World from Node
 test("Hello World from Node", async () => {
 	const response = await node.handle(ctx(), {});
-	expect({ message: "Hello World from Node!" }, response.data);
+	const message: ParamsDictionary = { message: "Hello World from Node!" };
+
+	expect(message).toEqual(response.data);
 });