+++ core/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nanoservice-ts/shared
 
+## 0.0.2
+
+### Patch Changes
+
+- Fixed open telemetry issues and types
+
 ## 0.0.1
 
 ### Patch Changes