+++ nodes/control-flow/if-else@1.0.0/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nanoservice-ts/if-else
 
+## 0.0.12
+
+### Patch Changes
+
+- Fixed open telemetry issues and types
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.8
+  - @nanoservice-ts/shared@0.0.2
+
 ## 0.0.11
 
 ### Patch Changes