+++ templates/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # node-name
 
+## 0.0.8
+
+### Patch Changes
+
+- Fixed open telemetry issues and types
+- Updated dependencies
+  - @nanoservice-ts/runner@0.1.8
+  - @nanoservice-ts/shared@0.0.2
+
 ## 0.0.7
 
 ### Patch Changes